+++ HISTORY.md
@@ -10,7 +10,7 @@
 * Fixed the logic of populating native data structure for `read_amp_bytes_per_bit` during OPTIONS file parsing on big-endian architecture. Without this fix, original code introduced in PR7659, when running on big-endian machine, can mistakenly store read_amp_bytes_per_bit (an uint32) in little endian format. Future access to `read_amp_bytes_per_bit` will give wrong values. Little endian architecture is not affected.
 
 ### Public API Change
-* Changed public but rarely-used FilterBitsBuilder::CalculateNumEntry to use size_t parameter and return.
+* Deprecated public but rarely-used FilterBitsBuilder::CalculateNumEntry, which is replaced with CalculateNumEntries taking a size_t parameter and returning size_t.
 
 ## 6.15.0 (11/13/2020)
 ### Bug Fixes